Commit code print report QPP011

+++ nts.uk/file/pr/nts.uk.file.pr.app/src/main/java/nts/uk/file/pr/app/export/residentialtax/data/ResidentialTaxSlipDto.java
@@ -1,5 +1,7 @@
 package nts.uk.file.pr.app.export.residentialtax.data;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import nts.arc.time.GeneralDate;
@@ -9,14 +11,14 @@ import nts.arc.time.GeneralDate;
 public class ResidentialTaxSlipDto {
 	private String resiTaxCode;
 	private int yearMonth;
-	private int taxPayrollMny;
-	private int taxRetirementMny;
-	private int taxOverdueMny;
-	private int taxDemandChargeMny;
+	private BigDecimal taxPayrollMny;
+	private BigDecimal taxRetirementMny;
+	private BigDecimal taxOverdueMny;
+	private BigDecimal taxDemandChargeMny;
 	private String address;
 	private GeneralDate dueDate;
-	private int headCount;
-	private int retirementBonusAmount;
-	private int cityTaxMny;
-	private int prefectureTaxMny;
+	private BigDecimal headCount;
+	private BigDecimal retirementBonusAmount;
+	private BigDecimal cityTaxMny;
+	private BigDecimal prefectureTaxMny;
 }
